Added method implemention
